fix: apply Living Globe to actual anonymous landing + delete dead code
The Living Globe changes were applied to the wrong component. The actual
anonymous homepage is components/hub/AnonymousLanding.tsx. Enable
interactive mode, breathing, health atmosphere, hover cards, narrative.
Delete 3,233 lines of dead legacy home components.
