+++ components/hub/AnonymousLanding.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState, useCallback } from 'react';
 import Link from 'next/link';
 import { ArrowRight, Users } from 'lucide-react';
 import { Button } from '@/components/ui/button';
@@ -8,6 +8,8 @@ import dynamic from 'next/dynamic';
 import { trackFunnel, FUNNEL_EVENTS } from '@/lib/funnel';
 import { useTranslation } from '@/lib/i18n/useTranslation';
 import { BarChart3 } from 'lucide-react';
+import { useQuery } from '@tanstack/react-query';
+import { motion, AnimatePresence, useReducedMotion } from 'framer-motion';
 import { GovernanceConsequenceCard } from './GovernanceConsequenceCard';
 import { IntelligencePreview } from './IntelligencePreview';
 import { ConversationalMatchFlow } from '@/components/matching/ConversationalMatchFlow';
@@ -15,6 +17,7 @@ import { CommunityPulse } from '@/components/intelligence/CommunityPulse';
 import { useFeatureFlag } from '@/components/FeatureGate';
 import { cn } from '@/lib/utils';
 import type { ConstellationRef } from '@/components/GovernanceConstellation';
+import type { ConstellationNode3D } from '@/lib/constellation/types';
 
 const ConstellationScene = dynamic(
   () => import('@/components/ConstellationScene').then((m) => ({ default: m.ConstellationScene })),
@@ -41,8 +44,22 @@ export function AnonymousLanding({ pulseData }: AnonymousLandingProps) {
   const { t } = useTranslation();
   const globeRef = useRef<ConstellationRef>(null);
   const [isMatching, setIsMatching] = useState(false);
+  const [hoveredNode, setHoveredNode] = useState<ConstellationNode3D | null>(null);
+  const prefersReducedMotion = useReducedMotion();
   const conversationalMatchingEnabled = useFeatureFlag('conversational_matching');
   const communityIntelligenceEnabled = useFeatureFlag('community_intelligence');
+
+  // Fetch dynamic governance narrative
+  const { data: narrativeData } = useQuery({
+    queryKey: ['homepage-narrative'],
+    queryFn: async () => {
+      const res = await fetch('/api/homepage/narrative');
+      if (!res.ok) return null;
+      return res.json() as Promise<{ narrative: string; healthScore: number; urgency: number }>;
+    },
+    staleTime: 5 * 60 * 1000,
+    refetchOnWindowFocus: false,
+  });
 
   useEffect(() => {
     trackFunnel(FUNNEL_EVENTS.LANDING_VIEWED);
@@ -52,6 +69,10 @@ export function AnonymousLanding({ pulseData }: AnonymousLandingProps) {
     setIsMatching(true);
     trackFunnel(FUNNEL_EVENTS.MATCH_STARTED, { source: 'landing_conversational' });
   };
+
+  const handleNodeHover = useCallback((node: ConstellationNode3D | null) => {
+    setHoveredNode(node);
+  }, []);
 
   return (
     <div className="relative flex flex-col min-h-[calc(100vh-4rem)]">
@@ -66,12 +87,70 @@ export function AnonymousLanding({ pulseData }: AnonymousLandingProps) {
         )}
       >
         <div className="absolute inset-0 overflow-hidden">
-          <ConstellationScene ref={globeRef} className="w-full h-full" interactive={false} />
+          <ConstellationScene
+            ref={globeRef}
+            className="w-full h-full"
+            interactive={!isMatching}
+            breathing={!prefersReducedMotion && !isMatching}
+            healthScore={narrativeData?.healthScore ?? 75}
+            urgency={narrativeData?.urgency ?? (pulseData?.activeProposals ?? 0) * 2}
+            onNodeHover={!isMatching ? handleNodeHover : undefined}
+          />
         </div>
 
         {/* Gradient fade — hidden when matching to keep immersion */}
         {!isMatching && (
           <div className="absolute inset-x-0 bottom-0 h-40 bg-gradient-to-t from-background to-transparent pointer-events-none" />
+        )}
+
+        {/* Hover card for globe nodes */}
+        <AnimatePresence>
+          {hoveredNode && !isMatching && (
+            <motion.div
+              initial={{ opacity: 0, y: 8 }}
+              animate={{ opacity: 1, y: 0 }}
+              exit={{ opacity: 0, y: 4 }}
+              transition={{ duration: 0.15 }}
+              className="pointer-events-none fixed bottom-20 left-1/2 -translate-x-1/2 z-50 rounded-xl border border-white/10 bg-black/85 backdrop-blur-md px-4 py-3 shadow-2xl max-w-[280px]"
+            >
+              <p className="text-sm font-semibold text-white truncate">
+                {hoveredNode.name || `${hoveredNode.id.slice(0, 12)}...`}
+              </p>
+              <div className="flex items-center gap-3 mt-1 text-xs text-white/60">
+                <span
+                  className={
+                    hoveredNode.nodeType === 'drep'
+                      ? 'text-teal-400'
+                      : hoveredNode.nodeType === 'spo'
+                        ? 'text-violet-400'
+                        : 'text-amber-400'
+                  }
+                >
+                  {hoveredNode.nodeType === 'drep'
+                    ? 'DRep'
+                    : hoveredNode.nodeType === 'spo'
+                      ? 'SPO'
+                      : 'CC'}
+                </span>
+                <span>
+                  Score <strong className="text-white/90">{hoveredNode.score}</strong>
+                </span>
+              </div>
+              <p className="text-[10px] text-white/40 mt-1">Click to explore</p>
+            </motion.div>
+          )}
+        </AnimatePresence>
+
+        {/* Dynamic governance narrative — below hero text, above match flow */}
+        {!isMatching && narrativeData?.narrative && (
+          <div className="absolute bottom-8 left-0 right-0 z-10 pointer-events-none flex justify-center px-6">
+            <p
+              className="text-xs sm:text-sm text-white/50 text-center max-w-md"
+              style={{ textShadow: '0 2px 12px rgba(0,0,0,0.8)' }}
+            >
+              {narrativeData.narrative}
+            </p>
+          </div>
         )}
 
         {/* Hero content */}
